+++ config/test.exs
@@ -7,6 +7,9 @@ config :ex_unit, capture_log: true
 
 config :commanded, event_store_adapter: Commanded.EventStore.Adapters.InMemory
 
+config :commanded, Commanded.EventStore.Adapters.InMemory,
+  serializer: Commanded.Serialization.JsonSerializer
+
 config :commanded_ecto_projections,
   ecto_repos: [Commanded.Projections.Repo],
   repo: Commanded.Projections.Repo